pyproject.toml
[tool.poetry]
name = "rofi_tpb"
version = "0.1.0"
description = ""
authors = ["Loic Coyle <loic.coyle@hotmail.fr>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
python-rofi = "^1.0.1"
tpblite = "^0.5"
lxml = "^4.5.2"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
rofi-tpb = 'rofi_tpb.cli:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
